requirements.txt
numpy<2.0.0
opencv-python>=4.8.0,<5.0.0
Pillow>=9.0.0
Flask>=2.3.0

# Optional dependencies for future WebRTC support
# aiortc>=1.5.0
# websockets>=11.0

# For ResNet50 and advanced ML models
# torch>=2.0.0
# torchvision>=0.15.0
# tensorflow>=2.13.0